--- a/src/ZstdSharp/FodyWeavers.xml
+++ b/src/ZstdSharp/FodyWeavers.xml
@@ -1,4 +1,3 @@
 ﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-  <InlineIL/>
   <InlineMethod />
 </Weavers>