--- a/src/ZstdSharp/FodyWeavers.xsd
+++ b/src/ZstdSharp/FodyWeavers.xsd
@@ -5,63 +5,6 @@
     <xs:complexType>
       <xs:all>
         <xs:element name="InlineMethod" minOccurs="0" maxOccurs="1" type="xs:anyType" />
-        <xs:element name="InlineIL" minOccurs="0" maxOccurs="1">
-          <xs:complexType>
-            <xs:attribute name="SequencePoints">
-              <xs:annotation>
-                <xs:documentation>Defines if sequence points should be generated for each emitted IL instruction. Default value: Debug</xs:documentation>
-              </xs:annotation>
-              <xs:simpleType>
-                <xs:restriction base="xs:string">
-                  <xs:enumeration value="Debug">
-                    <xs:annotation>
-                      <xs:documentation>Insert sequence points in Debug builds only (this is the default).</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                  <xs:enumeration value="Release">
-                    <xs:annotation>
-                      <xs:documentation>Insert sequence points in Release builds only.</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                  <xs:enumeration value="True">
-                    <xs:annotation>
-                      <xs:documentation>Always insert sequence points.</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                  <xs:enumeration value="False">
-                    <xs:annotation>
-                      <xs:documentation>Never insert sequence points.</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                </xs:restriction>
-              </xs:simpleType>
-            </xs:attribute>
-            <xs:attribute name="Warnings">
-              <xs:annotation>
-                <xs:documentation>Defines how warnings should be handled. Default value: Warnings</xs:documentation>
-              </xs:annotation>
-              <xs:simpleType>
-                <xs:restriction base="xs:string">
-                  <xs:enumeration value="Warnings">
-                    <xs:annotation>
-                      <xs:documentation>Emit build warnings (this is the default).</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                  <xs:enumeration value="Ignore">
-                    <xs:annotation>
-                      <xs:documentation>Do not emit warnings.</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                  <xs:enumeration value="Errors">
-                    <xs:annotation>
-                      <xs:documentation>Treat warnings as errors.</xs:documentation>
-                    </xs:annotation>
-                  </xs:enumeration>
-                </xs:restriction>
-              </xs:simpleType>
-            </xs:attribute>
-          </xs:complexType>
-        </xs:element>
       </xs:all>
       <xs:attribute name="VerifyAssembly" type="xs:boolean">
         <xs:annotation>