--- a/src/ZstdSharp/UnsafeHelper.cs
+++ b/src/ZstdSharp/UnsafeHelper.cs
@@ -91,7 +91,15 @@ namespace ZstdSharp
         public static T* GetArrayPointer<T>(T[] array) where T : unmanaged
         {
             var size = (uint)(sizeof(T) * array.Length);
+#if NET5_0_OR_GREATER
+            // This function is used to allocate memory for static data blocks.
+            // We have to use AllocateTypeAssociatedMemory and link the memory's
+            // lifetime to this assembly, in order to prevent memory leaks when
+            // loading the assembly in an unloadable AssemblyLoadContext.
+            var destination = (T*)RuntimeHelpers.AllocateTypeAssociatedMemory(typeof(UnsafeHelper), (int)size);
+#else
             var destination = (T*)malloc(size);
+#endif
             fixed (void* source = &array[0])
                 System.Runtime.CompilerServices.Unsafe.CopyBlockUnaligned(destination, source, size);