This PR:

* Removes the dependency to the `InlineIL` package. All its uses can be replaced by the `System.Runtime.CompilerServices.Unsafe` package or other built-in .NET APIs, and the replaced code is simpler.
* Updates the high-level `Compressor` and `Decompressor` classes to use classes derived from `SafeHandle` to manage the ZSTD context pointers. While safe handles are usually used for interop with native code, we use them here in managed code to manage the lifetime of manually allocated memory. This makes the code simpler and much more reliable; you don't have to write `GC.KeepAlive`s anymore, and safe handles protect from cases like one thread calling `Dispose` while another thread uses the resource.
* Removes two cases where spans were compared to `null`. This is not valid code since spans are structs and not classes.