@oleg-st 2024-07-08
1) The removal of the `InlineIL` and `InlineMethod` dependencies should depend on the benchmark results. I think `InlineIL` can be removed and if there are problems in older frameworks, we can leave the `InlineIL` dependency and the old implementation for them. It is more difficult to remove `InlineMethod`, because even .NET 8 benefits from its use. 
2) `SafeHandle` change LGTM overall with some comments.
3) I'm not removing support for older frameworks yet as long as it costs almost nothing to support them.