@teo-tsirpanis 2024-07-08
Oh you wrote `InlineMethod`, nice! And I hadn't seen that there are optimizations for all targeted frameworks.

One advantage to targeting less frameworks is reduced package size; right now it is in 1.15MB, which is not yet but close to what I would call "big". But it's not the end of the world.

All feedback was addressed, and I also fixed a memory leak I spotted. I am going to run some benchmarks and will get back to you.